+++ libs/MatLib/MatLib/sparse_matrix.cpp
@@ -58,8 +58,8 @@ namespace OpenMEEG {
 
         Tank::const_iterator it;
         for(it = m_tank.begin(); it != m_tank.end(); ++it) {
-            unsigned i = it->first.first;
-            unsigned j = it->first.second;
+            size_t i = it->first.first;
+            size_t j = it->first.second;
             double val = it->second;
             ret(i) += val * x(j);
         }
@@ -74,10 +74,10 @@ namespace OpenMEEG {
         out.set(0.0);
 
         for( Tank::const_iterator it = m_tank.begin(); it != m_tank.end(); ++it) {
-            unsigned i = it->first.first;
-            unsigned j = it->first.second;
+            size_t i = it->first.first;
+            size_t j = it->first.second;
             double val = it->second;
-            for ( unsigned k = 0; k < mat.ncol(); ++k) {
+            for ( size_t k = 0; k < mat.ncol(); ++k) {
                 out(i, k) += val * mat(j, k);
             }
         }
@@ -87,13 +87,13 @@ namespace OpenMEEG {
 
     SparseMatrix SparseMatrix::operator*(const SparseMatrix &mat) const
     {
-        // TODO: fast enough ?
+        // fast enough ?
         assert(ncol() == mat.nlin());
         SparseMatrix out(nlin(), mat.ncol());
 
         for ( Tank::const_iterator it1 = m_tank.begin(); it1 != m_tank.end(); ++it1) {
-            unsigned i = it1->first.first;
-            unsigned j = it1->first.second;
+            size_t i = it1->first.first;
+            size_t j = it1->first.second;
             for ( Tank::const_iterator it2 = mat.begin(); it2 != mat.end(); ++it2) {
                 if ( it2->first.first == j ) {
                     out(i, it2->first.second) += it1->second * it2->second;
@@ -109,13 +109,13 @@ namespace OpenMEEG {
         SparseMatrix out(nlin(), ncol());
 
         for ( Tank::const_iterator it = m_tank.begin(); it != m_tank.end(); ++it) {
-            unsigned i = it->first.first;
-            unsigned j = it->first.second;
+            size_t i = it->first.first;
+            size_t j = it->first.second;
             out(i, j) += it->second;
         }
         for ( Tank::const_iterator it = mat.begin(); it != mat.end(); ++it) {
-            unsigned i = it->first.first;
-            unsigned j = it->first.second;
+            size_t i = it->first.first;
+            size_t j = it->first.second;
             out(i, j) += it->second;
         }
         return out;
@@ -125,8 +125,8 @@ namespace OpenMEEG {
         SparseMatrix tsp(ncol(),nlin());
         const_iterator it;
         for(it = m_tank.begin(); it != m_tank.end(); ++it) {
-            unsigned i = it->first.first;
-            unsigned j = it->first.second;
+            size_t i = it->first.first;
+            size_t j = it->first.second;
             tsp(j,i) = it->second;
         }
         return tsp;
@@ -142,10 +142,10 @@ namespace OpenMEEG {
 
         double minv = m_tank.begin()->second;
         double maxv = m_tank.begin()->second;
-        unsigned mini = 0;
-        unsigned maxi = 0;
-        unsigned minj = 0;
-        unsigned maxj = 0;
+        size_t mini = 0;
+        size_t maxi = 0;
+        size_t minj = 0;
+        size_t maxj = 0;
 
         Tank::const_iterator it;
         for(it = m_tank.begin(); it != m_tank.end(); ++it) {
@@ -164,7 +164,7 @@ namespace OpenMEEG {
         std::cout << "Max Value : " << maxv << " (" << maxi << "," << maxj << ")" << std::endl;
         std::cout << "First Values" << std::endl;
 
-        unsigned cnt = 0;
+        size_t cnt = 0;
         for(it = m_tank.begin(); it != m_tank.end() && cnt < 5; ++it) {
             std::cout << "(" << it->first.first << "," << it->first.second << ") " << it->second << std::endl;
             cnt++;
